src/test/java/ru/netology/stats/StatsServiceTest.java
package ru.netology.stats;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.CsvSource;

public class StatsServiceTest {
    @Test
    void TestAverageMax() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int actual = service.amountTotalAverageMax(sales);
        int expected = 5;
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void TestAverageMin() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int actual = service.amountTotalAverageMin(sales);
        int expected = 5;
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void TestMinSales() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int actual = service.minSales(sales);
        int expected = 9;
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void TestMaxSales() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int actual = service.maxSales(sales);
        int expected = 8;
        Assertions.assertEquals(actual, expected);
    }

    @Test
    void TestTotalAverage() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int actual = service.amountTotalAverage(sales);
        int expected = 15;
        Assertions.assertEquals(actual, expected);
    }
    @Test
    void TestTotal() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int actual = service.amountTotal(sales);
        int expected = 180;
        Assertions.assertEquals(actual, expected);
    }

}